demo-restful/requirements.txt
Flask-HTTPAuth
passlib
itsdangerous
mongoengine